touch: remove the filter/prediction code
we aren't making any use of this, and while it's kind of nice in theory,
it's a lot of complexity to this code, and I'd like to reduce complexity
so that we might be able to autotest the handler going forward.

so just delete it all. it's always easy to re-add in a different way if
we ever need it.