touch: move more of the report code processing to reportPoints
this makes processInputEvent smaller and easier to reason about,
and also makes it easier to reason about the process of finally
creating touch points before reporting them.

it could be argued that it would be nice to split reportPoints into
two methods (e.g. handleSync which builds points), but hey ho.