touch: remove touchpoint storage
rather than persisting this as a member, and having a function to
populate the member, just build a list inline in reportPoints.
this makes it easier to see the flow of data here.

we also remove m_lastTouchPoints since that was unused in the present
code. previously, it was used by prediction/filtering, but if that is
reintroduced, it should be applied as a separate layer anyway.