touch: remove strange release report code
As far as I can determine, the idea here must be if a slot transitions
from one ID to another, make sure a release is sent. In practice, I
don't see how this can ever happen, since the full state is always kept
in m_contacts -- the only case I could ever imagine this breaking would
be if there's a mismatch between kernel state and the parsed state, but
I'd expect a lot more to break in this case.

Alternatively, maybe this is some kind of type A thing that has been
left lying around?