touch: add debug visualisation
to use, one can now run an application with EPTOUCH_DEBUG=foo.png myapp.
touch events are written (unfiltered) to an image buffer, and flushed to
the file on a timer.

the original implementation here is inspired by stream-touch-events --
I'm not sure who the original author of that was, but my thanks to them.
the only large difference here is that we render the rect of the
touchpoint, so we get a little more information from the kernel about
the size of the touch than that script gave us.