touch: handle missing releases correctly
in the case where a touchpoint is down, and then released while the
device is suspended, we won't get a release event reported. the same is
also (theoretically) true if we somehow manage to get out of sync with
the kernel state.

we now handle this by sending a release of the stuck point if it
happens.

this is an effective revert of cc6e2b04f5c8dedd0638e97043968ecb12497251,
but I had to rewrite it as the code has changed significantly since then.