fix(epaper-qpa): provide a QPlatformTheme, and override StartDragDistance
the default StartDragDistance Qt uses is 10, which is a very small
distance. this can manifest as missed presses when nesting e.g:

    Flickable { MouseArea { ... } }

because the Flickable will see a movement over StartDragDistance, and
hijack the press from the MouseArea.

by raising StartDragDistance, we raise the threshold for that stealing
to happen, making a missed press on a sloppy gesture less likely.

there's a small consideration, though: this also makes dragging
things (like selection handles) harder, so we shouldn't raise this
arbitrarily high without some care and consideration.

Task: SW-21396